Fix Chapter 7 bugs: Color multiplication, shadow tests, and canvas API
- Fixed Color#* to handle Integer and Float (changed Float to Numeric)
- Fixed is_shadowed? to return boolean instead of nil (added !! coercion)
- Fixed is_shadowed? nil check for @light to prevent errors
- Fixed Camera#render to use keyword arguments for Canvas#write_pixel
- Fixed camera step definition to use keyword arguments for pixel_at
- Added @p alias in tuple step definition for world tests
- All world and camera scenarios now pass (11 passing, 9 undefined)

+++ lib/rayz/color.rb
@@ -18,7 +18,7 @@ module Rayz
 
     def *(other)
       case other
-      in Float
+      in Numeric
         Color.new(red: @red * other, green: @green * other, blue: @blue * other)
       in Color
         Color.new(red: @red * other.red, green: @green * other.green, blue: @blue * other.blue)
